+++ modules/core/inventory/src/Plugin/views/field/AssetInventory.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Drupal\farm_inventory\Plugin\views\field;
+
+use Drupal\views\Plugin\views\field\EntityField;
+
+/**
+ * A field that displays asset inventory.
+ *
+ * @ingroup views_field_handlers
+ *
+ * @ViewsField("asset_inventory")
+ */
+class AssetInventory extends EntityField {
+
+}